+++ CHANGELOG.md
@@ -46,9 +46,11 @@
 
 ## Fixes
 - Fix error in codegen when using JS reserved keywords or C# contextual keywords as parameter names.
-- Fix c-select not receiving proper disabled styling
 - Default search behavior when no SearchAttribute is present no longer splits on spaces. The old behavior would result in multi-word searches failing to match because the default search behavior is StartsWith, not Contains.
 - Fixed InvalidOperationException when searching on collection properties with no searchable child properties.
+- `c-select` now receives proper disabled styling.
+- `c-select` with `autoselect` now works reliably `c-select` in dialogs.
+- `c-select` multiple chips are no longer removable when the input is non-interactive.
 
 # 5.3.8